Add user specific dashboards.

+++ smartlaundry-api/src/main/java/ke/co/smartlaundry/model/Role.java
@@ -8,9 +8,8 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    // ❌ was "role_id", but your schema defines column name as "id"
     @Column(name = "id")
-    private Short id; // SMALLSERIAL -> Short
+    private Long id;
 
     @Column(unique = true, nullable = false, length = 50)
     private String name; // e.g. ADMIN, OWNER, STAFF, DRIVER, CUSTOMER
@@ -21,8 +20,8 @@ public class Role {
         this.name = name;
     }
 
-    public Short getId() { return id; }
-    public void setId(Short id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
